Move serializers tests
